// src/types/FoodTruck.ts
import { Location } from "./Location";

export type FoodTruck = {
  objectid: string;
  applicant: string;
  facilitytype: string;
  cnn: string;
  locationdescription: string;
  address: string;
  blocklot: string;
  block: string;
  lot: string;
  permit: string;
  status: string;
  fooditems: string;
  x: string;
  y: string;
  latitude: string;
  longitude: string;
  schedule: string;
  received: string;
  priorpermit: string;
  expirationdate: Date;
  location: Location;
  ":@computed_region_yftq_j783": string;
  ":@computed_region_p5aj_wyqh": string;
  ":@computed_region_rxqg_mtj9": string;
  ":@computed_region_bh8s_q3mv": string;
  ":@computed_region_fyvs_ahh9": string;
  approved?: Date;
  dayshours: string;
  noisent?: Date;
};